backend/pyproject.toml
[tool.poetry]
name = "app-template"
version = "0.1.0"
description = ""
authors = ["peplxx"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
pydantic-settings = "^2.4.0"
uvicorn = "^0.30.5"
sqlalchemy = "^2.0.31"
pytz = "^2024.1"
fastapi-login = "^1.10.1"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
